Working Notification JsInterop Support
Fixed up the Notification.cs to support a notification with an empty
bitmap until we can workout what is causing the icons not to work
correctly.

Updated test project to pass a notification object to the index.html. In
the script tag of the index.html, a simple notification is sent to the
headset. Since it is transient, it disapears after a couple seconds.