Alpha and Attaching to Existing Overlays
Added an alpha property to the overlay object to control transparency.
Also added a new constructor to allow the passing of an existing overlay
key string which will get that overlay and load the object properties.
Hany when working with both your own overlays and others that may be
present.