Added Controller Events
Added support for the overlays to work when the controllers are turned
on after the application starts. Borrowed some code from the SteamVR
Unity Plugin to get the same behavior expected there.

It works by registering callbacks for the new device event and prevents
the browser from starting until it gets the work from the Overlay that
the controllers were found. Otherwise the overlay is created in memory
but the browser never starts up.