Controller Events (#8)
* Working Notification JsInterop Support

Fixed up the Notification.cs to support a notification with an empty
bitmap until we can workout what is causing the icons not to work
correctly.

Updated test project to pass a notification object to the index.html. In
the script tag of the index.html, a simple notification is sent to the
headset. Since it is transient, it disapears after a couple seconds.

* Now Passing Overlay Object to Notification Constructor Again

* Alpha and Attaching to Existing Overlays

Added an alpha property to the overlay object to control transparency.
Also added a new constructor to allow the passing of an existing overlay
key string which will get that overlay and load the object properties.
Hany when working with both your own overlays and others that may be
present.

* Added Controller Events

Added support for the overlays to work when the controllers are turned
on after the application starts. Borrowed some code from the SteamVR
Unity Plugin to get the same behavior expected there.

It works by registering callbacks for the new device event and prevents
the browser from starting until it gets the work from the Overlay that
the controllers were found. Otherwise the overlay is created in memory
but the browser never starts up.

* Incremental - Needs cleanup

Previous commit and pull request included a misunderstanding on my part
about how OpenVR was working, so I am in the middle of trying to fix
that now. SteamVR.cs did not need listeners embedded but instead they
belonged in SteamVR_WebKit.cs. By calling the SteamVR initialize to
attach the listeners, I was inadvertently preventing this from running
as an overlay only app. Other games launching were forcing it to quit.

Now that it's working again without the SteamVR.cs initialize being
called, I need to go through and cleanup SteamVR.cs. I imagine all of
the stuff I added last night could be ripped out.

* Cleanup Steam VR Listeners

Hand listeners for controller code in teh wrong place. Moved into
SteamVR_WebKit.cs and got rid of things that were not needed.

As a side note, I am starting to understand OpenVR a bit more. Still
feel like I'm bumping around in the dark looking for a light switch
though. At least something is working!

* Light Cleanup